US506882: Support JSONArray request - MASRequestBody support jsonArray - update test - changelog
